refactor: split event display headers
